refactor: extract stream service from core
