+++ src/api/Cargo.toml
@@ -23,6 +23,7 @@ enterprise = [
     "promql-service/enterprise",
     "search_service/enterprise",
     "enrichment-data/enterprise",
+    "stream/enterprise",
 ]
 cloud = [
     "enterprise",
@@ -30,12 +31,14 @@ cloud = [
     "openobserve-api-management/cloud",
     "openobserve-api-query/cloud",
     "openobserve-core/cloud",
+    "stream/cloud",
 ]
 vectorscan = [
     "openobserve-api-query/vectorscan",
     "openobserve-core/vectorscan",
     "promql-service/vectorscan",
     "search_service/vectorscan",
+    "stream/vectorscan",
 ]
 vortex = [
     "config/vortex",
@@ -62,6 +65,7 @@ audit.workspace = true
 db.workspace = true
 openobserve-core.workspace = true
 ingestion-common.workspace = true
+stream.workspace = true
 enrichment-data.workspace = true
 promql-service.workspace = true
 search.workspace = true
