+++ src/core/src/dashboards/mod.rs
@@ -36,8 +36,9 @@ use infra::{
         distinct_values::{DistinctFieldRecord, OriginType},
     },
 };
+use stream::save_stream_settings;
 
-use super::{db::distinct_values, folders, stream::save_stream_settings};
+use super::{db::distinct_values, folders};
 use crate::common::meta::authz::Authz;
 pub mod reports;
 pub mod timed_annotations;
