+++ src/jobs/src/job/cloud.rs
@@ -24,13 +24,11 @@ use config::{
 };
 use hashbrown::HashSet;
 use infra::table::org_users::get_admin;
+use stream::get_streams;
 
 use crate::{
     common::{infra::config::ORGANIZATIONS, meta::telemetry},
-    service::{
-        organization::is_org_in_free_trial_period, self_reporting::search::get_usage,
-        stream::get_streams,
-    },
+    service::{organization::is_org_in_free_trial_period, self_reporting::search::get_usage},
 };
 
 /// This file has all odd-jobs that are specific to cloud installation,
