+++ src/jobs/src/job/mod.rs
@@ -185,14 +185,9 @@ async fn enforce_usage_stream_retention() {
         && s.data_retention < 32
     {
         s.data_retention = 32;
-        openobserve_core::stream::save_stream_settings(
-            META_ORG_ID,
-            USAGE_STREAM,
-            StreamType::Logs,
-            s,
-        )
-        .await
-        .unwrap(); //unwrap is intentional, we should panic if this fails
+        stream::save_stream_settings(META_ORG_ID, USAGE_STREAM, StreamType::Logs, s)
+            .await
+            .unwrap(); //unwrap is intentional, we should panic if this fails
     }
 }
