+++ src/super_cluster_queue/Cargo.toml
@@ -12,8 +12,9 @@ enterprise = [
     "compaction/enterprise",
     "enrichment-data/enterprise",
     "openobserve-core/enterprise",
+    "stream/enterprise",
 ]
-vectorscan = ["o2_enterprise/vectorscan"]
+vectorscan = ["o2_enterprise/vectorscan", "stream/vectorscan"]
 
 [dependencies]
 anyhow.workspace = true
@@ -28,6 +29,7 @@ log.workspace = true
 o2_enterprise = { workspace = true, optional = true, default-features = false }
 openobserve-core.workspace = true
 compaction.workspace = true
+stream.workspace = true
 enrichment-data.workspace = true
 serde.workspace = true
 serde_json.workspace = true
