+++ src/super_cluster_queue/src/schemas.rs
@@ -130,9 +130,7 @@ async fn delete_fields(msg: Message) -> Result<()> {
 async fn stream_delete(msg: Message) -> Result<()> {
     let (org_id, stream_type, stream_name) = parse_key(&msg.key)?;
 
-    if let Err(e) =
-        openobserve_core::stream::stream_delete_inner(&org_id, stream_type, &stream_name).await
-    {
+    if let Err(e) = stream::stream_delete_inner(&org_id, stream_type, &stream_name).await {
         log::error!(
             "[SUPER_CLUSTER:sync] Failed to delete stream: {org_id}/{stream_type}/{stream_name}, error: {e}"
         );
